+++ sampler/__init__.py
@@ -1,9 +1,0 @@
-from .get_micro_batch import (slice_adj,SAGE,EdgeIndex,get_micro_batch)
-
-__all__ = [
-    'EdgeIndex',
-    'SAGE',
-    'get_micro_batch',
-]
-
-classes = __all__