+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [0.1.0] - 2023-MM-DD
 ### Added
+- Neighborloader with quiver_feature in quiver/minibatch_reddit_quiver_loader.py
 - Neighborloader Pyg2.2.0 sample examples.
 - Finish benchmark gloo. Benchmark sample time, Backward+Forward time and distributed broadcast + scatter time.
 - Import benchmark utils.py